* origin/Andi_dev:
  test new proximity_driver
  change proximity pins;
  Add measure error handling
  fine tune of measure speed
  red led out

# Conflicts:
#	A2Z-CAR-F7/A2Z-F7-Project/SW4STM32/STM32746G_DISCOVERY/Debug/STM32746G_DISCOVERY_led_matrix.bin
#	A2Z-CAR-F7/A2Z-F7-Project/SW4STM32/STM32746G_DISCOVERY/Debug/STM32746G_DISCOVERY_led_matrix.elf
#	A2Z-CAR-F7/A2Z-F7-Project/SW4STM32/STM32746G_DISCOVERY/Debug/output.map
#	Projects/A2Z-CAR/Inc/proximity_driver.h
#	Projects/A2Z-CAR/Src/main.c
#	Projects/A2Z-CAR/Src/wifi_functions.c

+++ A2Z-CAR-F7/A2Z-F7-Project/Src/socket_server.c
@@ -17,7 +17,7 @@
 typedef struct
 {
 	uint8_t buff_adc_data[9];
-	uint16_t buff_distance;
+	uint32_t buff_distance;
 } sensor_data;
 
 /* Private function prototypes -----------------------------------------------*/
